requirements.txt
numpy
scikit-learn
joblib
PyYAML

fastapi
uvicorn[standard]
pydantic
